refactor(background): extrai helpers de store para módulo compartilhado
storePublication/storeComment/storeEngagement/emptyMetrics/trackedHandleForProvider
saem do controller para src/background/store.ts, prontos para os módulos por-provider
importarem. Comportamento byte-idêntico (golden-master verde).

Refs #5
